Decoupled vulkan pipeline from device
